Introduce transformAsync with quarkus-managed worker threads + logging update
There is a new and consistent way to execute a transformation through the script-engine based on transformAsync that takes a TransformJob and outputs a TransformationResult object. The entire process is logged with DEBUG/ERROR/INFO Levels for easier bugtracking.

One working Example is testable through GET http://localhost:8091/api/scripts/trigger-job

Current setup is natively compilable and requires a RabbitMQ docker container running.