ConsumerControlManagement Implementation and Tests finished
Awaits user actions related messages of the stay-in-sync-core. Supported Categories are the starting, updating, and deleting of a supported pollingJob.

The class also rejects messages with obviously faulty syncJobs (fields == null, apiAddressSet.isEmpty(), negative values).

The tests should check for the correct message handeling, as well as the blocking of faulty messages.
The tests haven´t been run yet, because of other unfinished classes blocking the process