Add Resources for Transformation and TransformationScript
We now have api endpoints to use crud operations on Transformation and TransformationScript entities.

A Transformation is a composite that gets created as a shell initially and then associated with scripts, rules, endpoints etc. in order to fully describe a Transformation

The pattern to use is [POST] with TrasnformationShellDTO, then [PUT] with TransformationAssemblyDTO.

Also check swagger-ui for example payloads and parameters.

TransformationScript have basic crud operations but require additional features like hash creation and validation.